+++ avocado-trees-ordering/place-order-handler/placeOrderHandler.js
@@ -2,7 +2,7 @@ var handlerName = 'place_chicken_order';
 var notifyELK = require('../../notifications/elk-notification/elkNotification');
 var createTranslator = require('../../utils/translator/translator');
 var translations = require('../translations');
-var translate =  createTranslator(translations, project.vars.cor_lang);
+var translate =  createTranslator(translations, service.vars.lang);
 var avocadoEligibility = require('../avocado-eligibility/avocadoEligibility');
 
 function isOrderEligible(input){