+++ just-in-time/justInTime.test.js
@@ -81,6 +81,19 @@ describe('clientRegistration', () => {
     });
     describe('Account number handler success callback', () => {
         var callback;
+        var orders = [];
+        var maizeOrder,secondOrder;
+        orders.push(mockBundleInput[0]);
+        maizeOrder = JSON.parse(JSON.stringify(mockBundleInputs[0]));
+        maizeOrder.bundleName = '0.5 Maize Acre';
+        maizeOrder.price = 4950;
+        maizeOrder.quantity = 0.5;
+        orders.push(maizeOrder);
+        secondOrder = JSON.parse(JSON.stringify(mockBundleInputs[0]));
+        secondOrder.bundleName = '0.25 Maize Acre';
+        secondOrder.price = 2830;
+        secondOrder.quantity = 0.25;
+        orders.push(secondOrder);
         const mockCursor = { next: jest.fn(), 
             hasNext: jest.fn()
         };
@@ -106,6 +119,20 @@ describe('clientRegistration', () => {
             expect(state.vars.main_menu).toEqual(`Select a product\n1) ${mockRow.vars.bundle_name}`+
             ` ${mockRow.vars.price}`+
             '\n');
+        });
+        it('should display the maize bundle size(0.5 and 0.25) of maize bundles if a maize bundle is available and the client did not order anything',()=>{
+            mockCursor.hasNext.mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(false).mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRows[0]).mockReturnValueOnce(mockRows[0]);//.mockReturnValueOnce(mockMaizeRows[0]).mockReturnValueOnce(mockMaizeRows[1]);
+            state.vars.orders = ' ';
+            callback();
+            expect(state.vars.bundles).toEqual(JSON.stringify(orders));
+        });
+        it('should display the maize bundle size(0.5 and 0.25) of maize bundles if a maize bundle is available and the client ordered other bundles than maize',()=>{
+            mockCursor.hasNext.mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(false).mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRows[0]).mockReturnValueOnce(mockRows[0]);//.mockReturnValueOnce(mockMaizeRows[0]).mockReturnValueOnce(mockMaizeRows[1]);
+            state.vars.orders = JSON.stringify(mockRows[1]);
+            callback();
+            expect(state.vars.bundles).toEqual(JSON.stringify(orders));
         });
         it('should display a the bundles with next and previous options if the prepayment condition is satified and there is a lot of bundles',()=>{
     
@@ -149,40 +176,110 @@ describe('clientRegistration', () => {
         it('should display a message saying that the prepayment condition is not satified if the remaining loan is greater than 500', () => {
             client.BalanceHistory[0].TotalCredit = 5000;
             client.BalanceHistory[0].TotalRepayment_IncludingOverpayments = 0;
-            var amount = 4500;
+            var amount = 500;
             state.vars.topUpClient = JSON.stringify(client);
             callback();
             expect(sayText).toHaveBeenCalledWith(`You do not qualify for a top up, pay at least ${amount}`+
             ' to qualify.');
         });
+        
     });
     describe('bundle Choice Handler successfull callback',()=>{
         var callback;
+        const mockCursor = { next: jest.fn(), 
+            hasNext: jest.fn()
+        };
+        const mockRow = {vars: {'bundleId': '-2009','bundleInputId': '-1709','bundle_name': 'Second possible name bundle','price': '2251','input_name': 'second input'}};
+        beforeAll(() => {
+            const mockTable = { queryRows: jest.fn()};
+            mockTable.queryRows.mockReturnValue(mockCursor);
+            project.initDataTableById.mockReturnValue(mockTable);
+        });
         beforeEach(() => {
             justInTime.registerHandlers();
             callback = bundleChoiceHandler.getHandler.mock.calls[0][0];    
-            state.vars.bundleInputs =  JSON.stringify(mockBundleInputs);      
+            state.vars.chosenMaizeBundle =  ' ';      
         });
 
         it('should display the varieties for the bundle choosed, if the bundle has multiple varieties',()=>{
+            state.vars.varietyBundleId = -2009;
+            mockCursor.hasNext.mockReturnValueOnce(true).mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow);
             callback(mockBundleInputs[0].bundleId);
             expect(sayText).toHaveBeenCalledWith(`Select seed variety\n1) ${mockBundleInputs[0].inputName}`+
             `\n2) ${mockBundleInputs[3].inputName}`+
             '\n');
         });
+        it('should display a the varieties for the bundle choosed with next and previous options if there is a lot of varieties for that bundle',()=>{
+            state.vars.varietyBundleId = -2009;
+            mockCursor.hasNext.mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(true).mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow).mockReturnValueOnce(mockRow);
+            callback(mockBundleInputs[0].bundleId);
+            //expect(sayText).not.toHaveBeenCalledWith(expect.stringContaining('You do not qualify for a top up,'));
+            expect(state.vars.main_menu).toEqual(`Select seed variety\n1) ${mockRow.vars.input_name}`+
+            `\n2) ${mockRow.vars.input_name}`+
+            `\n3) ${mockRow.vars.input_name}`+
+            `\n4) ${mockRow.vars.input_name}`+
+            `\n5) ${mockRow.vars.input_name}`+
+            `\n6) ${mockRow.vars.input_name}`+
+            `\n7) ${mockRow.vars.input_name}`+
+            '\n77)Next page');
+        });
+        
         it('should display the order placed by the user and prompt to add order or finalize order',()=>{
-            state.vars.bundleInputs =  JSON.stringify(mockBundleInput);  
+            mockCursor.hasNext.mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow); 
             state.vars.orders = ' ';
-            var orderMessage = mockBundleInput[0].bundleName + ' ' + mockBundleInput[0].price + '\n';
-            callback(mockBundleInput[0].bundleId);
+            var orderMessage = mockBundleInputs[0].bundleName + ' ' + mockBundleInputs[0].price + '\n';
+            callback(mockBundleInputs[0].bundleId);
             expect(sayText).toHaveBeenCalledWith(`Order placed\n ${orderMessage}`+
-            ' \n 1) Add product\n 2) Finish ordering');
+            ' \n 1) Add product\n 2) Finish ordering\n 3) Go back');
             expect(promptDigits).toHaveBeenCalledWith(addOrderHandler.handlerName);
         });
-
+        
+        it('should add the chosen bundle to the ordered bundle list',()=>{
+            mockCursor.hasNext.mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow); 
+            var newList = mockBundleInputs;
+            state.vars.orders = JSON.stringify(mockBundleInputs);
+            newList.push(mockBundleInputs[0]);
+            callback(mockBundleInputs[0].bundleId);
+            expect(state.vars.orders).toEqual(JSON.stringify(newList));
+        });
+        it('should show the correct order price if the maize bundle is chosen',()=>{
+            var maizeBundle =  mockRow.vars;
+            maizeBundle.bundleName = '0.5 Maize';
+            maizeBundle.price = 750;
+            state.vars.orders = ' ';
+            state.vars.chosenMaizeBundle = JSON.stringify(maizeBundle);
+            mockCursor.hasNext.mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow); 
+            callback(maizeBundle.bundleId);
+            var orderMessage = maizeBundle.bundleName + ' ' + maizeBundle.price + '\n';
+            expect(sayText).toHaveBeenCalledWith(`Order placed\n ${orderMessage}`+
+            ' \n 1) Add product\n 2) Finish ordering\n 3) Go back');
+        }); 
+        it('should display the final message with the bundles if the number of ordered bundles is 3',()=>{
+            var firstBundles = [{'bundleId': '-2009','bundleInputId': '-1709','bundleName': 'Second possible name bundle','price': '2251','inputName': 'second input'},{'bundleId': '-9009','bundleInputId': '-5709','bundleName': 'third possible name bundle','price': '6251','inputName': 'third input'}];
+            var newBundle = {'bundleId': '-8004','bundleInputId': '-8709','bundleName': 'fourth possible name bundle','price': '5251','inputName': 'fourth input'};
+            var mockRow = {vars: {'bundleId': '-8004','bundleInputId': '-8709','bundle_name': 'fourth possible name bundle','price': '5251','input_name': 'fourth input'}};
+            state.vars.orders = JSON.stringify(firstBundles);
+            var allBundles =  [{'bundleId': '-2009','bundleInputId': '-1709','bundleName': 'Second possible name bundle','price': '2251','inputName': 'second input'},{'bundleId': '-9009','bundleInputId': '-5709','bundleName': 'third possible name bundle','price': '6251','inputName': 'third input'}, {'bundleId': '-8004','bundleInputId': '-8709','bundleName': 'fourth possible name bundle','price': '5251','inputName': 'fourth input'}];
+            mockCursor.hasNext.mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow); 
+            callback(newBundle.bundleId);
+            var orderMessage = '';
+            for( var j = 0; j < allBundles.length; j++ ){
+                orderMessage = orderMessage + allBundles[j].bundleName + ' ' + allBundles[j].price + '\n';
+            }
+            expect(sayText).toHaveBeenCalledWith(`Order placed\n ${orderMessage}`+
+            ' \n1) Confirm order\n 2) Go back');
+        });  
     });
     describe('order Confirmation Handler successfull callback',()=>{
         var callback;
+        const mockTable = { queryRows: jest.fn(), createRow: jest.fn() };
+        const mockRow = {save: jest.fn()};
         var orders = [{'bundleName': 'Maize','price': 1000},{'bundleName': 'Pesticide','price': 3000}];
         var ordersMessage = orders[0].bundleName + ' ' + orders[0].price + ' '+orders[1].bundleName + ' ' + orders[1].price+' ';
         beforeAll(()=>{
@@ -196,18 +293,22 @@ describe('clientRegistration', () => {
         it('should send a message confirming the order is complete if the order is saved in roster',()=>{
             httpClient.request.mockReturnValue({status: 201});
             contact.phone_number = phoneNumber;
-            var message = `Thanks for ordering ${ordersMessage}`+
-            '. Thank you for topping-up through JiT.';
+            mockTable.createRow.mockReturnValueOnce(mockRow);
+            project.initDataTableById.mockReturnValue(mockTable);
+            var message = `Thank you for topping-up through JiT. Your order is ${ordersMessage}`+
+            ' Reach out to CE through *689# if this  order is not correct';
             callback();
             expect(sayText).toHaveBeenCalledWith(message);
             expect(project.sendMessage).toHaveBeenCalledWith({content: message, to_number: phoneNumber});
         });
         it('should send a message to the stored client\'s phone confirming the order is complete if the order is saved in roster',()=>{
+            mockTable.createRow.mockReturnValueOnce(mockRow);
+            project.initDataTableById.mockReturnValue(mockTable);
             var inactive_number = {'PhoneNumber': '0786182098', 'IsInactive': false};
             var active_number ={'PhoneNumber': '0786182098', 'IsInactive': true};
             getPhoneNumber.mockReturnValue([inactive_number, active_number]);
-            var message = `Thanks for ordering ${ordersMessage}`+
-            '. Thank you for topping-up through JiT.';
+            var message = `Thank you for topping-up through JiT. Your order is ${ordersMessage}`+
+            ' Reach out to CE through *689# if this  order is not correct';
             callback();
             expect(sayText).toHaveBeenCalledWith(message);
             expect(project.sendMessage).toHaveBeenCalledWith({content: message, to_number: active_number.PhoneNumber});
@@ -237,6 +338,15 @@ describe('clientRegistration', () => {
  
     });
     describe('variety Confirmation Handler successfull callback',()=>{
+        const mockCursor = { next: jest.fn(), 
+            hasNext: jest.fn()
+        };
+        const mockRow = {vars: {'bundleId': '-2009','bundleInputId': '-1709','bundle_name': 'Second possible name bundle','price': '2251','input_name': 'second input'}};
+        beforeAll(() => {
+            const mockTable = { queryRows: jest.fn()};
+            mockTable.queryRows.mockReturnValue(mockCursor);
+            project.initDataTableById.mockReturnValue(mockTable);
+        });
         var callback;
         beforeEach(() => {
             justInTime.registerHandlers();
@@ -245,26 +355,40 @@ describe('clientRegistration', () => {
             state.vars.bundleInputs =  JSON.stringify(mockBundleInputs);        
         });
         it('should show the order placed',()=>{
+            state.vars.chosenMaizeBundle = ' ';
+            state.vars.varietyBundleId = -2009;
+            mockCursor.hasNext.mockReturnValueOnce(true);
+            mockCursor.next.mockReturnValueOnce(mockRow);
             callback(mockBundleInputs[3].bundleId,true,mockBundleInputs[3].bundleInputId);
             var orderMessage = mockBundleInputs[3].bundleName + ' ' + mockBundleInputs[3].price + '\n';
             expect(sayText).toHaveBeenCalledWith(`Order placed\n ${orderMessage}`+
-            ' \n 1) Add product\n 2) Finish ordering');
+            ' \n 1) Add product\n 2) Finish ordering\n 3) Go back');
 
         });
-
     });
     describe('add Order Handler successfull callback',()=>{
         var callback;
         beforeEach(() => {
             justInTime.registerHandlers();
             state.vars.orders = JSON.stringify(mockBundleInput);
-            callback = addOrderHandler.getHandler.mock.calls[0][0];          
+            callback = addOrderHandler.getHandler.mock.calls[0][0];   
+              
         });
         it('should show the order placed',()=>{
             callback();
             var orderMessage = mockBundleInput[0].bundleName + ' ' + mockBundleInput[0].price + '\n';
             expect(sayText).toHaveBeenCalledWith(`Order placed\n ${orderMessage}`+
-            ' \n1) Confirm order');
+            ' \n1) Confirm order\n 2) Go back');
+        });
+        it('should show the correct order price if the maize bundle is chosen',()=>{
+            var maizeBundle =  mockBundleInput[0];
+            maizeBundle.bundleName = '0.5 Maize';
+            maizeBundle.price = 750;
+            state.vars.chosenMaizeBundle = JSON.stringify(maizeBundle);
+            callback();
+            var orderMessage = maizeBundle.bundleName + ' ' + maizeBundle.price + '\n';
+            expect(sayText).toHaveBeenCalledWith(`Order placed\n ${orderMessage}`+
+            ' \n1) Confirm order\n 2) Go back');
         });
     });