+++ rw-legacy/core.js
@@ -63,6 +63,9 @@ var client_table = project.initDataTableById(service.vars['21a_client_data_id'])
 // console.log(JSON.stringify(service.vars));
 // console.log('### End Env Info###');
 
+var lang = service.vars.lang || project.vars.cor_lang;
+service.vars.lang = lang;
+
 var slack = require('../slack-logger/index');
 var notifyELK = require('../notifications/elk-notification/elkNotification');
 //global functionss
@@ -85,7 +88,6 @@ var checkGroupLeader = require('../shared/rosterApi/checkForGroupLeader');
 var avocadoTreesOrdering = require('../avocado-trees-ordering/avocadoTreesOrdering');
 var clientRegistration = require('../client-registration/clientRegistration');
 //options
-const lang = project.vars.cor_lang;
 const max_digits_for_input = project.vars.max_digits; //only for testing
 //const max_digits_for_nid = parseInt(settings_table.queryRows({'vars' : {'settings' : 'max_digits_nid'}}).next().vars.value); 
 const max_digits_for_account_number = project.vars.max_digits_an;
@@ -149,6 +151,15 @@ addInputHandler('account_number_splash', function (input) { //acount_number_spla
             sayText(current_menu);
             promptDigits('enr_reg_start', { 'submitOnHash': false, 'maxDigits': max_digits_for_nid, 'timeout': timeout_length });
         }
+    } else if(response == 99) { // use 99 to change the language
+        lang = service.vars.lang == 'ki' ? 'en' : 'ki'
+        service.vars.lang = lang;
+        sayText(msgs('cor_enr_main_splash',{},lang));
+        promptDigits('account_number_splash', {
+            'submitOnHash': false,
+            'maxDigits': 2,
+            'timeout': timeout_length
+        });
     }
     else {
         try {